handout/requirements.txt
# Core
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
pandas>=1.5.0

# nuScenes
nuscenes-devkit>=1.1.9
pyquaternion>=0.9.9

# Visualization
matplotlib>=3.6.0
open3d>=0.16.0

# Utilities
tqdm>=4.64.0
Pillow>=9.0.0
